requirements.txt
flask
langchain-core
langchain-ollama
langgraph
pydantic
typing-extensions
stripe
python-dotenv
requests